+++ packages/widget-dashboard/src/components/widget-actions/widget-actions.tsx
@@ -1,16 +1,12 @@
-import { privateApis as componentsPrivateApis } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { moreVertical } from '@wordpress/icons';
-// eslint-disable-next-line @wordpress/use-recommended-components
-import { Icon, IconButton, Link } from '@wordpress/ui';
+// eslint-disable-next-line @wordpress/use-recommended-components -- Intentional early adoption of the new Menu, pending WordPress/gutenberg#76135.
+import { Icon, IconButton, Menu } from '@wordpress/ui';
 import { useWidgetHost } from '@wordpress/widget-primitives';
 import type { WidgetAction } from '@wordpress/widget-primitives';
 import { getActionRoute } from './get-action-route';
 import { useReserveHeaderSpace } from '../widget-header/widget-header-fit';
 import styles from './widget-actions.module.css';
-import { unlock } from '../../lock-unlock';
-
-const { Menu } = unlock( componentsPrivateApis );
 
 type WidgetActionsProps = {
 	/**
@@ -48,8 +44,8 @@ export function WidgetActions( {
 
 	return (
 		<span ref={ reserveRef } className={ styles[ 'widget-actions' ] }>
-			<Menu>
-				<Menu.TriggerButton
+			<Menu.Root>
+				<Menu.Trigger
 					render={
 						<IconButton
 							icon={ moreVertical }
@@ -61,55 +57,39 @@ export function WidgetActions( {
 					}
 				/>
 
-				<Menu.Popover>
-					<Menu.Group className={ styles[ 'widget-action-items' ] }>
+				<Menu.Popup>
+					<Menu.Group>
 						{ actions.map( ( action ) => {
 							const path = getActionRoute( links, action );
 							const HostLink = links?.Link;
+							const linkProps =
+								path !== null && HostLink
+									? { render: <HostLink path={ path } /> }
+									: {
+											href: action.href,
+											download: action.download,
+											openInNewTab: action.openInNewTab,
+									  };
 
 							return (
-								<Menu.Item
+								<Menu.LinkItem
 									key={ action.id }
+									{ ...linkProps }
 									prefix={
 										action.icon ? (
 											<Icon icon={ action.icon } />
 										) : undefined
 									}
-									render={
-										path !== null && HostLink ? (
-											<Link
-												className={
-													styles[
-														'widget-action-link'
-													]
-												}
-												render={
-													<HostLink path={ path } />
-												}
-											/>
-										) : (
-											<Link
-												href={ action.href }
-												download={ action.download }
-												openInNewTab={
-													action.openInNewTab
-												}
-												className={
-													styles[
-														'widget-action-link'
-													]
-												}
-											/>
-										)
-									}
 								>
-									{ action.label }
-								</Menu.Item>
+									<Menu.ItemLabel>
+										{ action.label }
+									</Menu.ItemLabel>
+								</Menu.LinkItem>
 							);
 						} ) }
 					</Menu.Group>
-				</Menu.Popover>
-			</Menu>
+				</Menu.Popup>
+			</Menu.Root>
 		</span>
 	);
 }